+++ doc/source/conf.py
@@ -26,7 +26,7 @@ author = 'Andrew Yoon'
 # The short X.Y version
 version = ''
 # The full version, including alpha/beta/rc tags
-release = '0.0.19-dev'
+release = '0.0.19'
 
 
 # -- General configuration ---------------------------------------------------